Fixed bug with tcp frag in hybi-10 & pattern mismatch on socket close
hybi-10 was randomly failing when tcp fragmentation happened due to secondary passes through i_handle_data
some times accidentally matching the remaining payload as a valid header.
This caused a corrupt read and subsequent crash of the socket process.

also fixed is the handling of {websocket_close,Data}, was re-using an already bound variable (Data) causing a pattern
match fail in cases where the websocket module was trying to close a socket.

Changed to use a new CloseData variable to allow the match to happen and the response to be sent